Split the check logic to a separated file.
